createWood and getter methods for block settings

+++ src/main/java/creoii/creo/common/item/BoatItem.java
@@ -41,7 +41,7 @@ public class BoatItem extends Item {
                 Vec3d vec3d2 = user.getEyePos();
 
                 for (Entity entity : list) {
-                    Box box = entity.getBoundingBox().expand((double) entity.getTargetingMargin());
+                    Box box = entity.getBoundingBox().expand(entity.getTargetingMargin());
                     if (box.contains(vec3d2)) {
                         return TypedActionResult.pass(itemStack);
                     }
